pyproject.toml
[tool.poetry]
name = "gpt-api-interface"
version = "0.1.0"
description = ""
authors = ["Matthias Osswald <matthias@busykoala.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.14.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
